+++ Lib/fontbakery/specifications/adobe_fonts.py
@@ -68,7 +68,7 @@ expected_check_ids = [
     'com.google.fonts/check/079',  # Monospace font has hhea.advanceWidthMax equal to each glyph's advanceWidth?
     'com.google.fonts/check/152',  # Name table strings must not contain 'Reserved Font Name'.
     'com.google.fonts/check/family_and_style_max_length',
-    'com.google.fonts/check/167',  # The variable font 'wght' (Weight) axis coordinate must be 400 on the 'Regular' instance.
+    'com.google.fonts/check/varfont/regular_wght_coord',
     'com.google.fonts/check/168',  # The variable font 'wdth' (Width) axis coordinate must be 100 on the 'Regular' instance.
     'com.google.fonts/check/169',  # The variable font 'slnt' (Slant) axis coordinate must be zero on the 'Regular' instance.
     'com.google.fonts/check/170',  # The variable font 'ital' (Italic) axis coordinate must be zero on the 'Regular' instance.